+++ tests/machine/x/rust/substring_builtin.rs
@@ -1,3 +1,3 @@
 fn main() {
-    println!("{}", &"mochi"[1 as usize..4 as usize]);
+    println!("{}", &"mochi"[1..4]);
 }
